feat: prompt user with evaluation modal on drag to kanban finish section
